requirements.txt
absl-py==0.6.1
appnope==0.1.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
docutils==0.15.2
entrypoints==0.3
gast==0.2.1
gensim==3.8.1
grpcio==1.17.1
h5py==2.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
more-itertools==5.0.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.6
numpy==1.16.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==4.1.0
pytest-runner==4.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.22.0
s3transfer==0.3.2
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
smart-open==1.9.0
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-probability==0.5.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.2
tqdm==4.29.0
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
metis
